asn1: add 'bitset' field tag
X.690 11.2.2 says that 'Where [named bit list], applies, the
bitstring shall have all trailing 0 bits removed before it is
encoded.'

The x509 code currently does this manually; this commit adds
a new 'bitset' field tag to generalize this.  The tag indicates
that a BIT STRING is a collection of named bits and so should
have trailing zeros removed.