x509: replace NonFatalErrors with Errors
Convert the x509 parsing code that currently populates a NonFatalErrors
structure for non-fatal errors to instead populate a more detailed
Errors structure, including more details about what part of the RFC
the error pertains to.

The Errors structure was previously only used for the revocation structure
processing, and was specific to this fork of crypto/x509.